fixing admin to deploy

+++ Backend/main/Admin/Admin/src/main/java/csd/backend/Admin/Repository/TournamentPlayerRepository.java
@@ -6,7 +6,7 @@ import csd.backend.Admin.Model.Tournament.Tournament;
 import csd.backend.Admin.Model.Tournament.TournamentPlayer;
 import java.util.*;
 
-public interface TournamentPlayerRepository extends JpaRepository<TournamentPlayer, Integer>{
+public interface TournamentPlayerRepository extends JpaRepository<TournamentPlayer, Long>{
     Optional<TournamentPlayer> findByPlayerId(Long playerId);
     List<TournamentPlayer> findByTournament(Tournament tournament);
     Optional<TournamentPlayer> findByPlayerIdAndTournamentId(Long playerId, Long tournamentId);
