+++ Backend/main/Admin/Admin/src/main/java/csd/backend/Admin/Repository/TournamentRoundRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import csd.backend.Admin.Model.Tournament.Tournament;
 import csd.backend.Admin.Model.Tournament.TournamentRound;
 
-public interface TournamentRoundRepository extends JpaRepository<TournamentRound, Integer>{
+public interface TournamentRoundRepository extends JpaRepository<TournamentRound, Long>{
     // Find the latest round by tournament (highest round number)
     Optional<TournamentRound> findTopByTournamentOrderByRoundNumberDesc(Tournament tournament);
 }
